+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## [0.0.1] - TODO: Add release date.
+## [0.1.2] - Added support for invisible moneybutton
 
-* TODO: Describe initial release.
+## [0.0.1] - An experimental moneybutton plugin for flutter web
+
+